Fix delete course and group

+++ manage_courses.php
@@ -12,6 +12,19 @@ $result = mysqli_query($link,"SELECT * FROM courses");
 
 ?>
 <div class="content">
+
+<?php
+// Display success/error message
+if (isset($_SESSION['message'])){
+  echo "<font color='blue'>".$_SESSION['message']."</font>";
+  unset($_SESSION['message']);
+}
+if (isset($_SESSION['error'])){
+    echo "<font color='red'>".$_SESSION['error']."</font>";
+    unset($_SESSION['error']);
+}
+?>
+
 <h1>Manage Courses</h1>
 </br>
 <div  class='form-group'>
@@ -44,7 +57,7 @@ echo "<td>" . $row['term'] . "</td>";
 echo "<td>" . $row['year'] . "</td>";
 echo "<td>" . $row['start_date'] . "</td>";
 echo "<td>" . $row['end_date'] . "</td>";
-echo "<td><a href='edit_course.php?id=".$row['course_id']."'>Edit</a>/<a href='delete_course.php?id=".$row['course_id']."'>Delete</a>/<a href='manage_sections.php?id=".$row['course_id']."'>Sections</a></td>";
+echo "<td><a href='edit_course.php?id=".$row['course_id']."'>Edit</a>/<a href='delete_course.php?id=".$row['course_id']."' onclick=\"return confirm('Are you sure you want to delete this course?')\">Delete</a>/<a href='manage_sections.php?id=".$row['course_id']."'>Sections</a></td>";
 echo "</tr>";
 }
 echo "</table>";
